Housekeeping: separeted custom_feedback model
