+++ lib/src/modules/feedback/models/models.dart
@@ -1,0 +1,1 @@
+export 'custom_feedback.dart';
